Fix bug when a macro injects its text into the tty.
This function, tty_pushstring(), was not written properly,
yet it worked for 25 years across many platforms.
I call tty_insert_flip_string, and that in turn calls memcpy,
which is a simple routine except it is optimized in assembly for each platform,
so not that simple.
In all its variations, it was ok copying data from user space
into kernel space, which is what I was doing.
Somewhere in kernel 6, memcpy was modified further, and now blows up
when copying from user space into kernel space.
So this change writes it the way it should have been written in the first place.
I mean, I use get_user all over the place, so I was aware
of the user kernel crossover issues, I just didn't put it together
for this function.